+++ src/commands/hardis/org/select.ts
@@ -2,12 +2,7 @@
 import { flags, SfdxCommand } from "@salesforce/command";
 import { Messages } from "@salesforce/core";
 import { AnyJson } from "@salesforce/ts-types";
-import * as c from "chalk";
-import { MetadataUtils } from "../../../common/metadata-utils";
-import { execSfdxJson, uxLog } from "../../../common/utils";
-import { prompts } from "../../../common/utils/prompts";
-import { WebSocketClient } from "../../../common/websocketClient";
-import { setConfig } from "../../../config";
+import { promptOrg } from "../../../common/utils/orgUtils";
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -58,64 +53,9 @@ export default class OrgSelect extends SfdxCommand {
     const devHub = this.flags.devhub || false;
     this.debugMode = this.flags.debug || false;
 
-    // List all local orgs and request to user
-    const orgListResult = await MetadataUtils.listLocalOrgs("any");
-    const orgList = [
-      ...(orgListResult?.scratchOrgs || []),
-      ...(orgListResult?.nonScratchOrgs || []),
-      { username: "Connect to another org", otherOrg: true },
-      { username: "Cancel", cancel: true },
-    ];
-    const orgResponse = await prompts({
-      type: "select",
-      name: "org",
-      message: c.cyanBright("Please select an org"),
-      choices: orgList.map((org: any) => {
-        const title = org.username || org.alias || org.instanceUrl;
-        const description = title !== org.instanceUrl ? org.instanceUrl : "";
-        return {
-          title: c.cyan(title),
-          description: description,
-          value: org,
-        };
-      }),
-    });
+    const org = await promptOrg(this, { devHub: devHub, setDefault: true });
 
-    // Cancel
-    if (orgResponse.org.cancel === true) {
-      uxLog(this, c.cyan("Cancelled"));
-      process.exit(0);
-    }
-
-    // Connect to new org
-    if (orgResponse.org.otherOrg === true) {
-      await this.config.runHook("auth", {
-        checkAuth: true,
-        Command: this,
-        devHub,
-      });
-      return { outputString: "Launched org connection" };
-    } else {
-      // Set default username
-      const setDefaultUsernameCommand = `sfdx config:set ${devHub ? "defaultdevhubusername" : "defaultusername"}=${orgResponse.org.username}`;
-      await execSfdxJson(setDefaultUsernameCommand, this, {
-        fail: true,
-        output: false,
-      });
-      WebSocketClient.sendMessage({ event: "refreshStatus" });
-      // Update local user .sfdx-hardis.yml file with response if scratch has been selected
-      if (orgResponse.org.username.includes("scratch")) {
-        await setConfig("user", {
-          scratchOrgAlias: orgResponse.org.username,
-          scratchOrgUsername: orgResponse.org.alias || orgResponse.org.username,
-        });
-      }
-
-      uxLog(this, c.gray(JSON.stringify(orgResponse.org, null, 2)));
-      uxLog(this, c.cyan(`Selected org ${c.green(orgResponse.org.username)} - ${c.green(orgResponse.org.instanceUrl)}`));
-
-      // Return an object to be displayed with --json
-      return { outputString: `Selected org ${orgResponse.org.username}` };
-    }
+    // Return an object to be displayed with --json
+    return { outputString: `Selected org ${org.username}` };
   }
 }