+++ src/commands/hardis/project/configure/auth.ts
@@ -40,6 +40,7 @@ export default class ConfigureAuth extends SfdxCommand {
   };
 
   // Comment this out if your command does not require an org username
+  protected static supportsUsername = true;
   protected static requiresUsername = false;
 
   // Comment this out if your command does not support a hub org username
@@ -47,6 +48,8 @@ export default class ConfigureAuth extends SfdxCommand {
 
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = false;
+
+  protected static requiresDependencies = ['openssl'];
   /* jscpd:ignore-end */
 
   public async run(): Promise<AnyJson> {
@@ -100,7 +103,7 @@ export default class ConfigureAuth extends SfdxCommand {
     // Generate SSL certificate (requires openssl to be installed on computer)
     const certFolder = devHub ? "./config/.jwt" : "./config/branches/.jwt";
     const certName = devHub ? config.devHubAlias : branchName;
-    await generateSSLCertificate(certName, certFolder, this);
+    await generateSSLCertificate(certName, certFolder, this, this.org?.getConnection());
     // Return an object to be displayed with --json
     return { outputString: "Configured branch for authentication" };
   }